test(stylexswc/transform): put the canonicalization cases on one macro
The media query canonicalization file used both `stylex_test` and
`stylex_test_transform`, so half its cases carried their expected output inline
and half kept it in a snapshot. The file's own module comment describes the
snapshot as where the contract lives, and `stylex_test` is what the rest of the
suite reaches for, so the three inline cases move onto it.

The generated snapshots were checked line for line against the expectations
they replace and are identical, so nothing about what these tests assert has
changed. The corpus was rescanned because it is harvested from these files.

Also aligns the crate glossary's decision record links with the form used in
the other crates, which is the path rather than the words ADR 0001, so that
grepping for a record finds every reference to it.
