+++ crates/stylex-transform/tests/__swc_snapshots__/tests/transform_stylex_create_test/media_query_canonicalization.rs/a_disjoint_breakpoint_ladder_keeps_its_contradictory_branches.js
@@ -1,0 +1,39 @@
+import _inject from "@stylexjs/stylex/lib/stylex-inject";
+var _inject2 = _inject;
+import "colors.stylex.js";
+import * as stylex from '@stylexjs/stylex';
+import { colors } from 'colors.stylex.js';
+_inject2({
+    ltr: ".x1fm9ujy{color:var(--x1g366na)}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media ((not all) or (not all)) or ((not all) or ((min-width: 1440px))){.x11g08g8.x11g08g8{color:var(--x1gey9a0)}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (not all) or ((min-width: 1200px) and (max-width: 1439px)){.x1qsezja.x1qsezja{color:var(--xo0k5im)}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (min-width: 1024px) and (max-width: 1199px){.xqw8h0p.xqw8h0p{color:var(--x1ncemq0)}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (min-width: 768px) and (max-width: 1023px){.x1bhj7sf.x1bhj7sf{color:var(--x5eudzp)}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (min-width: 480px) and (max-width: 767px){.x1p9ejzw.x1p9ejzw{color:var(--xtqmhjj)}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (max-width: 479px){.xw70vyp.xw70vyp{color:var(--xuk8yok)}}",
+    priority: 3200
+});
+export const styles = {
+    root: {
+        kMwMTN: "x1fm9ujy x11g08g8 x1qsezja xqw8h0p x1bhj7sf x1p9ejzw xw70vyp",
+        $$css: true
+    }
+};
