+++ crates/stylex-transform/tests/__swc_snapshots__/tests/transform_stylex_create_test/media_query_canonicalization.rs/a_rewritten_media_key_sorts_where_the_authored_one_did.js
@@ -1,0 +1,39 @@
+import _inject from "@stylexjs/stylex/lib/stylex-inject";
+var _inject2 = _inject;
+import * as stylex from '@stylexjs/stylex';
+_inject2({
+    ltr: ".x7z7khe{padding:10px}",
+    priority: 1000
+});
+_inject2({
+    ltr: ".x1mqxbix{color:black}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@supports (display: grid){.x19g4ih5.x19g4ih5{color:green}}",
+    priority: 3030
+});
+_inject2({
+    ltr: "@container (min-width: 400px){.x15pkjp4.x15pkjp4{color:teal}}",
+    priority: 3300
+});
+_inject2({
+    ltr: "@media not all{.x1jqaanj.x1jqaanj{color:red}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (min-width: 100px){.x18tmubq.x18tmubq{color:blue}}",
+    priority: 3200
+});
+_inject2({
+    ltr: ".xy3p2pi{margin:2px}",
+    priority: 1000
+});
+export const styles = {
+    root: {
+        kmVPX3: "x7z7khe",
+        kMwMTN: "x1mqxbix x19g4ih5 x15pkjp4 x1jqaanj x18tmubq",
+        kogj98: "xy3p2pi",
+        $$css: true
+    }
+};
