+++ crates/stylex-transform/tests/__swc_snapshots__/tests/transform_stylex_create_test/media_query_canonicalization.rs/colliding_rewritten_keys_drop_a_declaration.js
@@ -1,0 +1,25 @@
+import _inject from "@stylexjs/stylex/lib/stylex-inject";
+var _inject2 = _inject;
+import * as stylex from '@stylexjs/stylex';
+_inject2({
+    ltr: ".x1mqxbix{color:black}",
+    priority: 3000
+});
+_inject2({
+    ltr: "@media not all{.x12vud9h.x12vud9h{color:blue}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (max-width: 99.99px) and (min-height: 100px){.xsllcrx.xsllcrx{color:green}}",
+    priority: 3200
+});
+_inject2({
+    ltr: "@media (min-width: 100px){.xr6za1w.xr6za1w{color:purple}}",
+    priority: 3200
+});
+export const styles = {
+    root: {
+        kMwMTN: "x1mqxbix x12vud9h xsllcrx xr6za1w",
+        $$css: true
+    }
+};
